infra/export undo-redo from editor-common (#1880)
