+++ CHANGELOG.md
@@ -23,6 +23,10 @@
 - `mobileNativeLoader`
   - [#1867](https://github.com/wix-incubator/rich-content/pull/1867) create mobile native plugin loader bundle
 
+### :house: Internal
+- `editor-common`
+  - [#1880](https://github.com/wix-incubator/rich-content/pull/1880) moved `handleUndoRedoCommands` to editor-common
+
 </details>
 <hr/>
