+++ bundle-analyzer/bundlesSizesBaseline.json
@@ -33,9 +33,9 @@
   "ricos-viewer-no-modal": 168,
   "ricos-viewer-with-plugins": 196,
   "rce": 679,
-  "rce-with-plugins": 850,
+  "rce-with-plugins": 849,
   "ricos-editor": 704,
   "ricos-editor-no-modal": 704,
-  "ricos-editor-with-plugins": 875,
+  "ricos-editor-with-plugins": 874,
   "ricos-common": 16
 }
