+++ packages/editor-common/web/src/index.ts
@@ -33,6 +33,7 @@ export { default as RichContentModal } from './Modals/RichContentModal';
 export { default as decorateComponentWithProps } from './Utils/decorateComponentWithProps';
 export { getToolbarTheme } from './Utils/getToolbarTheme';
 export { getModalStyles, getBottomToolbarModalStyles } from './Utils/getModalStyles';
+export { undo, redo } from './Utils/handleUndoRedoCommands';
 
 export {
   updateLinkAtCurrentSelection,
