+++ packages/plugin-undo-redo/web/src/RedoButton.jsx
@@ -2,8 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import redoIcon from './icons/RedoIcon';
-import { InlineToolbarButton } from 'wix-rich-content-editor-common';
-import { redo } from './utils';
+import { InlineToolbarButton, redo } from 'wix-rich-content-editor-common';
 
 const RedoButton = props => {
   const {
