+++ packages/plugin-undo-redo/web/src/insert-buttons.tsx
@@ -1,4 +1,10 @@
-import { TOOLBARS, INSERT_PLUGIN_BUTTONS, BUTTON_TYPES } from 'wix-rich-content-editor-common';
+import {
+  TOOLBARS,
+  INSERT_PLUGIN_BUTTONS,
+  BUTTON_TYPES,
+  undo,
+  redo,
+} from 'wix-rich-content-editor-common';
 import UndoIcon from './icons/UndoIcon';
 import RedoIcon from './icons/RedoIcon';
 import {
@@ -8,7 +14,6 @@ import {
   SetEditorState,
 } from 'wix-rich-content-common';
 import { UndoRedoPluginEditorConfig } from './types';
-import { undo, redo } from './utils';
 
 const createInsertButtons: CreateInsertButtons = ({
   t,
